+++ CHANGELOG.md
@@ -1,5 +1,44 @@
 # Changelog
 
++ **1.2.0** - _02/14/2018_
+  - Add the capability to indicate an endpoint param is required with the `required` key in the options hash.
+  ```
+        params.valid :message, :text,
+                     required: true,
+                     info: "the message of the post"
+  ```
+  - Add support for specifying the type of an endpoint param. For an endpoint param that has type `Array`,
+    type of list items can be specified using `item_type` key in the options hash.
+  ```
+        params.valid :viewable_by, :array,
+                     item_type: :integer,
+                     info: "an array of user ids that can access the post"
+  ```
+  - Add support for specifying the data type of an item for a presenter field using `item_type` key in the
+    options hash when the field is of type `Array`.
+  ```
+        field :aliases, :array,
+              item_type: :string,
+              info: "an array of user ids that can access the post"
+  ```
+  - Include the type and item type when generating markdown documentation for endpoint params.
+  - Specify the data type of a filter and available values with `items` key in the options hash. If filter is an array,
+    data type of items can be specified with the `item_type` property in options.
+  ```
+        filter :status, :string,
+               items: ['Started', 'Completed'],
+               info: "only returns elements with the given status"
+
+        filter :sprocket_ids, :array,
+               item_type: :integer,
+               info: "returns objects associated with given sprocket Ids"
+  ```
+  - Add support for generating markdown documentation for the following:
+    - when the `required` option is specified on an endpoint param
+    - when the `type` and `item_type` params are specified on the endpoint param
+    - when the `type` and `item_type` params are specified on a presenter field
+    - when the `type` and `items` params are specified on a presenter filter
+
 + **1.1.1** - _01/15/2017_
   - Add `Brainstem.mysql_use_calc_found_rows` boolean config option to utilize MySQL's [FOUND_ROWS()](https://dev.mysql.com/doc/refman/5.7/en/information-functions.html#function_found-rows) functionality to avoid issuing a new query to calculate the record count, which has the potential to up to double the response time of the endpoint.